moved and renamed entity_detail file according to hierarcy
