+++ lib/screens/home_FAB.dart
@@ -6,7 +6,7 @@ import 'package:iit_app/model/appConstants.dart';
 import 'package:iit_app/model/built_post.dart';
 import 'package:iit_app/model/colorConstants.dart';
 import 'package:iit_app/pages/council/councilPage.dart';
-import 'package:iit_app/ui/entity_detail.dart';
+import 'package:iit_app/pages/club_entity/entityPage.dart';
 
 Widget homeFAB(context, {fabKey}) =>
     AppConstants.councilsSummaryfromDatabase == null
