+++ lib/ui/club_council_common/club_&_council_widgets.dart
@@ -4,10 +4,10 @@ import 'package:iit_app/external_libraries/spin_kit.dart';
 import 'package:iit_app/model/appConstants.dart';
 import 'package:iit_app/model/built_post.dart';
 import 'package:iit_app/model/colorConstants.dart';
-import 'package:iit_app/pages/club/clubPage.dart';
+import 'package:iit_app/pages/club_entity/clubPage.dart';
 import 'package:iit_app/screens/create.dart';
 import 'package:iit_app/ui/club_council_common/description.dart';
-import 'package:iit_app/ui/entity_detail.dart';
+import 'package:iit_app/pages/club_entity/entityPage.dart';
 import 'package:iit_app/ui/separator.dart';
 import 'package:iit_app/ui/text_style.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
