+++ lib/ui/drawer.dart
@@ -3,7 +3,7 @@ import 'package:iit_app/model/appConstants.dart';
 import 'package:iit_app/model/built_post.dart';
 import 'package:iit_app/model/colorConstants.dart';
 import 'package:iit_app/ui/dialogBoxes.dart';
-import 'package:iit_app/ui/entity_detail.dart';
+import 'package:iit_app/pages/club_entity/entityPage.dart';
 import 'package:iit_app/ui/text_style.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:built_collection/built_collection.dart';
